refactor: introduce root cmd file and global flags
